[APDV-111] Moved files
