+++ backend/src/main/java/pl/edu/agh/apdvbackend/use_cases/field_parser/SaveNewFieldParserImpl.java
@@ -2,7 +2,7 @@ package pl.edu.agh.apdvbackend.use_cases.field_parser;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import pl.edu.agh.apdvbackend.mappers.FieldParserMapper;
+import pl.edu.agh.apdvbackend.mappers.field_parser.FieldParserMapper;
 import pl.edu.agh.apdvbackend.models.body_models.field_parser.AddFieldParserRequestBody;
 import pl.edu.agh.apdvbackend.models.database.FieldParser;
 import pl.edu.agh.apdvbackend.repositories.FieldParserRepository;
